fix: cargo sweep install for newer rust versions
Currently EarthBuild's rust helper functions install the most recent
release of cargo-sweep. However this release does not work with Rust
versions >= 1.85.0. So functionality was added to install cargo-sweep
directly from a commit that fixed the observed bug. This function
should be updated to install the next release of cargo-sweep once it's
published.